LPS-55310 migrate to module
