+++ modules/portal/portal-messaging/src/com/liferay/portal/messaging/internal/jmx/DestinationStatisticsManagerMBean.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.messaging.jmx;
+package com.liferay.portal.messaging.internal.jmx;
 
 /**
  * @author Michael C. Han
